Crate serde_spanned

A serde-compatible spanned Value

This allows capturing the location, in bytes, for a value in the original parsed document for compatible deserializers.

Modules

de

Deserialization support for Spanned

Structs

Spanned

A spanned value, indicating the range at which it is defined in the source.